requirements.txt
django
coreapi
django-filter
djangorestframework
mysqlclient
django-simple-captcha